+++ src/lib/config.ts
@@ -46,9 +46,9 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 
 export const SDK_METADATA = {
   language: "typescript",
-  openapiDocVersion: "4.17.1-alpha.1773270186742-a6cfcb14",
-  sdkVersion: "0.7.0-alpha.1",
-  genVersion: "2.856.1",
+  openapiDocVersion: "4.17.1-alpha.1773356600656-ad22ad35",
+  sdkVersion: "0.7.0-alpha.2",
+  genVersion: "2.859.2",
   userAgent:
-    "speakeasy-sdk/typescript 0.7.0-alpha.1 2.856.1 4.17.1-alpha.1773270186742-a6cfcb14 cribl-control-plane",
+    "speakeasy-sdk/typescript 0.7.0-alpha.2 2.859.2 4.17.1-alpha.1773356600656-ad22ad35 cribl-control-plane",
 } as const;