+++ RELEASES.md
@@ -1329,3 +1329,13 @@ Based on:
 - [typescript v0.7.0-alpha.1] .
 ### Releases
 - [NPM v0.7.0-alpha.1] https://www.npmjs.com/package/cribl-control-plane/v/0.7.0-alpha.1 - .
+
+## 2026-03-13 02:12:34
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.752.0 (2.859.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.7.0-alpha.2] .
+### Releases
+- [NPM v0.7.0-alpha.2] https://www.npmjs.com/package/cribl-control-plane/v/0.7.0-alpha.2 - .